requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
marshmallow
flask-marshmallow